refactor: 修改用户实体字段

+++ src/main/java/com/github/lingkai5wu/loveta/mapper/MenuMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public interface MenuMapper extends BaseMapper<Menu> {
 
-    List<Menu> listMenusByUserId(long id);
+    List<Menu> listMenusByUserId(int id);
 
     List<Menu> listMenusByRoleId(int id);
