+++ src/main/java/com/github/lingkai5wu/loveta/mapper/PermissionMapper.java
@@ -15,7 +15,7 @@ public interface PermissionMapper extends BaseMapper<Permission> {
 
     List<String> listPermissionCodesByUserId(Object id);
 
-    List<Permission> listPermissionsByUserId(long id);
+    List<Permission> listPermissionsByUserId(int id);
 
     List<Permission> listPermissionsByRoleId(int id);
