+++ src/main/java/com/github/lingkai5wu/loveta/mapper/RoleMapper.java
@@ -15,5 +15,5 @@ public interface RoleMapper extends BaseMapper<Role> {
 
     List<String> listRoleCodesByUserId(Object id);
 
-    List<Role> listRolesByUserId(long id);
+    List<Role> listRolesByUserId(int id);
 }
