+++ src/main/java/com/github/lingkai5wu/loveta/service/impl/PermissionServiceImpl.java
@@ -32,7 +32,7 @@ public class PermissionServiceImpl extends ServiceImpl<PermissionMapper, Permiss
     }
 
     @Override
-    public List<Permission> listPermissionsByUserId(long id) {
+    public List<Permission> listPermissionsByUserId(int id) {
         return baseMapper.listPermissionsByUserId(id);
     }
